requirements.txt
langchain
langchain-openai
langchain-community
azure-search-documents
pypdf
python-dotenv
azure-identity
fastapi
uvicorn[standard]
python-multipart
gunicorn